Moved output writing package to vibe-cli module.
