+++ vibe-cli/src/main/java/org/molgenis/vibe/cli/output/ValuesSeparator.java
@@ -1,4 +1,4 @@
-package org.molgenis.vibe.io.output;
+package org.molgenis.vibe.cli.output;
 
 /**
  * Defines possible {@code separators} to be used.
