+++ vibe-cli/src/main/java/org/molgenis/vibe/cli/output/format/gene_prioritized/GenePrioritizedOutputFormatWriterFactory.java
@@ -1,16 +1,17 @@
-package org.molgenis.vibe.io.output.format.gene_prioritized;
+package org.molgenis.vibe.cli.output.format.gene_prioritized;
 
 
+import org.molgenis.vibe.cli.output.ValuesSeparator;
+import org.molgenis.vibe.cli.output.format.OutputFormatWriter;
+import org.molgenis.vibe.cli.output.format.PrioritizedOutputFormatWriter;
+import org.molgenis.vibe.cli.output.format.PrioritizedOutputFormatWriterFactory;
+import org.molgenis.vibe.cli.output.target.OutputWriter;
 import org.molgenis.vibe.formats.Gene;
 import org.molgenis.vibe.formats.GeneDiseaseCollection;
-import org.molgenis.vibe.io.output.format.OutputFormatWriter;
-import org.molgenis.vibe.io.output.ValuesSeparator;
-import org.molgenis.vibe.io.output.format.PrioritizedOutputFormatWriterFactory;
-import org.molgenis.vibe.io.output.target.OutputWriter;
 import org.molgenis.vibe.query_output_digestion.prioritization.Prioritizer;
 
 /**
- * A factory for {@link Gene} prioritized {@link org.molgenis.vibe.io.output.format.PrioritizedOutputFormatWriter}{@code s}.
+ * A factory for {@link Gene} prioritized {@link PrioritizedOutputFormatWriter}{@code s}.
  */
 public enum GenePrioritizedOutputFormatWriterFactory implements PrioritizedOutputFormatWriterFactory<Gene> {
     SIMPLE {
