+++ vibe-cli/src/main/java/org/molgenis/vibe/cli/output/format/gene_prioritized/ResultsPerGeneSeparatedValuesOutputFormatWriterUsingUris.java
@@ -1,8 +1,8 @@
-package org.molgenis.vibe.io.output.format.gene_prioritized;
+package org.molgenis.vibe.cli.output.format.gene_prioritized;
 
 import org.molgenis.vibe.formats.*;
-import org.molgenis.vibe.io.output.ValuesSeparator;
-import org.molgenis.vibe.io.output.target.OutputWriter;
+import org.molgenis.vibe.cli.output.ValuesSeparator;
+import org.molgenis.vibe.cli.output.target.OutputWriter;
 import org.molgenis.vibe.query_output_digestion.prioritization.Prioritizer;
 
 import java.net.URI;
