+++ vibe-cli/src/test/java/org/molgenis/vibe/cli/CommandLineOptionsParserTest.java
@@ -6,9 +6,9 @@ import org.apache.commons.lang3.ArrayUtils;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.molgenis.vibe.formats.Phenotype;
-import org.molgenis.vibe.io.output.format.gene_prioritized.GenePrioritizedOutputFormatWriterFactory;
-import org.molgenis.vibe.io.output.target.FileOutputWriter;
-import org.molgenis.vibe.io.output.target.StdoutOutputWriter;
+import org.molgenis.vibe.cli.output.format.gene_prioritized.GenePrioritizedOutputFormatWriterFactory;
+import org.molgenis.vibe.cli.output.target.FileOutputWriter;
+import org.molgenis.vibe.cli.output.target.StdoutOutputWriter;
 import org.molgenis.vibe.ontology_processing.PhenotypesRetrieverFactory;
 
 import java.io.IOException;
