+++ vibe-cli/src/test/java/org/molgenis/vibe/cli/output/format/gene_prioritized/GenePrioritizedOutputFormatWriterFactoryTest.java
@@ -1,10 +1,10 @@
-package org.molgenis.vibe.io.output.format.gene_prioritized;
+package org.molgenis.vibe.cli.output.format.gene_prioritized;
 
 import org.junit.jupiter.api.*;
 import org.mockito.Mockito;
 import org.molgenis.vibe.formats.*;
-import org.molgenis.vibe.io.output.format.OutputFormatWriter;
-import org.molgenis.vibe.io.output.target.StdoutOutputWriter;
+import org.molgenis.vibe.cli.output.format.OutputFormatWriter;
+import org.molgenis.vibe.cli.output.target.StdoutOutputWriter;
 import org.molgenis.vibe.query_output_digestion.prioritization.Prioritizer;
 import org.molgenis.vibe.query_output_digestion.prioritization.gene.HighestSingleDisgenetScoreGenePrioritizer;
